Update price tracking logic and enhance PriceTracker component
- Changed the timePeriod parsing in buildMatchCriteria from integer to float to support fractional hours.
- Added a new dropdown option for "Last 30 Minutes" in the PriceTracker component to improve user selection for price drop tracking.

+++ backend/src/utils/priceTracking.js
@@ -46,7 +46,7 @@ export const buildMatchCriteria = (timePeriod, notUpdated, searchQuery) => {
 
   // Handle time period for price drops
   if (timePeriod && timePeriod !== "all") {
-    const hours = parseInt(timePeriod);
+    const hours = parseFloat(timePeriod);
     if (!isNaN(hours) && hours > 0) {
       const timeAgo = new Date(Date.now() - hours * 60 * 60 * 1000);
       criteria.priceDroppedAt = {
